+++ photos/seePhoto/seePhoto.resolvers.js
@@ -1,0 +1,12 @@
+import { transformDocument } from "@prisma/client/runtime";
+import client from "../../client";
+export default {
+  Query: {
+    seePhoto: (_, { id }) =>
+      client.photo.findUnique({
+        where: {
+          id,
+        },
+      }),
+  },
+};