+++ photos/uploadPhoto/uploadPhoto.typeDefs.js
@@ -1,0 +1,8 @@
+import { gql } from "apollo-server-express";
+
+export default gql`
+  scalar Upload
+  type Mutation {
+    uploadPhoto(file: String!, caption: String): Photo
+  }
+`;