+++ users/unfollowUser/unfollow.resolvers.js
@@ -1,0 +1,36 @@
+import client from "../../client";
+import { protectedResolver } from "../users.utils";
+
+export default {
+  Mutation: {
+    unfollowUser: protectedResolver(
+      async (_, { username }, { loggedInUser }) => {
+        const ok = await client.user.findUnique({ where: { username } });
+        if (!ok) {
+          return {
+            ok: false,
+            error: "Can't unfollow user.",
+          };
+        }
+
+        // 외부키로 연결된 following, followed은 connect, disconnet를 통해서 설정을 할 수 있음
+        await client.user.update({
+          where: {
+            id: loggedInUser.id,
+          },
+          data: {
+            following: {
+              disconnect: {
+                username,
+              },
+            },
+          },
+        });
+
+        return {
+          ok: true,
+        };
+      }
+    ),
+  },
+};