+++ users/users.resolvers.js
@@ -1,1 +1,60 @@
+import client from "../client";
 
+export default {
+  User: {
+    totalFollowing: ({ id }) =>
+      client.user.count({
+        where: {
+          followers: {
+            some: {
+              id,
+            },
+          },
+        },
+      }),
+    totalFollowers: ({ id }) =>
+      client.user.count({
+        where: {
+          following: {
+            some: {
+              id,
+            },
+          },
+        },
+      }),
+    isMe: ({ id }, _, { loggedInUser }) => {
+      if (!loggedInUser) {
+        return false;
+      }
+
+      return id == loggedInUser.id;
+    },
+    isFollowing: async ({ id }, _, { loggedInUser }) => {
+      if (!loggedInUser) {
+        return false;
+      }
+
+      //   const exists = await client.user
+      //     .findUnique({ where: { username: loggedInUser.username } })
+      //     .following({
+      //       where: {
+      //         id,
+      //       },
+      //     });
+
+      const exists = await client.user.count({
+        where: {
+          username: loggedInUser,
+          following: {
+            some: {
+              id,
+            },
+          },
+        },
+      });
+
+      //존재하면 length가 0이상이므로 true가 됨
+      return Boolean(exists);
+    },
+  },
+};