+++ users/users.typeDefs.js
@@ -1,15 +1,24 @@
 import { gql } from "apollo-server";
 
 export default gql`
-    type User {
-        id: String!
-        firstName: String!
-        lastName: String
-        username: String!
-        email: String!
-        createdAt: String!
-        updatedAt: String!
-        bio: String
-        avatar: String
-    }
+  type User {
+    id: Int!
+    firstName: String!
+    lastName: String
+    username: String!
+    email: String!
+    createdAt: String!
+    updatedAt: String!
+    bio: String
+    avatar: String
+    following: [User]
+    followers: [User]
+    totalFollowing: Int!
+    totalFollowers: Int!
+    isMe: Boolean!
+    isFollowing: Boolean!
+  }
 `;
+
+//
+//