+++ history.md
@@ -152,14 +152,15 @@
     원인은 아래 코드인데 getUser 함수에 await(비동기처리) 를 하지 않아서
     promise{null} 형태로 값이 발생됨(중요)
 
-    코드)
-    const server = new ApolloServer({
-    schema,
-    context: async ({ req }) => {
-    return {
-    loggedInUser: await getUser(req.headers.token),
-    };
-    },
+         코드)
+         const server = new ApolloServer({
+         schema,
+         context: async ({ req }) => {
+            return {
+               loggedInUser: await getUser(req.headers.token),
+            };
+         },
+
     });
 
 2.  파일 업로드
@@ -187,3 +188,98 @@
             readStream.pipe(writeStrema) // readStream 과 writeStream 연결 => 파일 저장됨
 
             //aws에 파일을 저장할 경우
+
+# 12/24
+
+    - followUser는 db의 user 테이블은 follower(follow 대상자)와 following(follow 주체자) 가 각자 복수로 존재해야한다.
+      그래서 prisma에서는 아래와 같이 관계명을 동일하게 "FollowRelation"으로 정하고 각 record에 유니크한 키 즉, id를
+      외래키로 정해서 followers 와 following이 서로 동기화 될 수 있도록 한다.
+
+        user 테이블에 아래와 같이 등록)
+
+        followers User[] @relation("FollowRelation", references: [id])
+        following User[] @relation("FollowRelation", references: [id])
+
+    - followUser resolver을 실행하면 로그인된 id와 following 대상자의 username을 받아와서 user 테이블에 로그인된 id로 user 정보를 찾고
+      해당 정보에 속해있는 following에 username(follow 대상이 되는)을 등록한다.
+
+# 12/25
+
+1.  User typeDefes에 followers, following 을 Array형으로 추가 (아래 참조)
+
+    예시)
+    following: [User]
+    followers: [User]
+
+2.  prisma에서 following, followers 처럼 Array유형으로 되어있는 경우 쿼리시
+    include 구문을 통해서 조회 될 수 있도록 처리해줘야함
+
+    예시)
+
+    ```
+      client.user.user.findUnique({
+          where: {
+            username,
+         },
+         include : {
+            following: true,
+            followers: trun
+         }
+      })
+
+    ```
+
+3.  pagination(많은 데이터를 나눠서 보여주는(쿼리하는) 기능)
+
+    - page 매개변수를 통해서 skip(스킵할 레코드 수), take(가져올 레코드 수) 자동으로
+      설정될 수 있도록 쿼리를 만든다.
+
+      예제)
+
+      ```
+      const followers = await client.user
+      .findUnique({ where: { username } })
+        .followers({
+           take: 5,
+           skip: (page - 1) \* 5,
+        });
+
+      ```
+
+    - lastId 매개변수를 통해서 cursor를 만드는 방식이다. 다음페이지는 skip, row는 고정한 상태에서
+      현재 보여준 페이지의 lastId를 cursor를 등록하면서 한 페이지씩 넘어가는 구조이다.
+      cursor 방식은 무한 스크롤를 구현할때 편하지만, 특정 페이지로 이동하지는 못하는 단점이 있다.
+
+      예제)
+
+      ```
+       const following = await client.user
+        .findUnique({ where: { username } })
+        .following({
+          take: 5,
+          skip: lastId ? 1 : 0,
+          ...(lastId && { cursor: { id: lastId } }),
+        });
+
+      ```
+
+# 12/26
+
+1.  Computed Field : 어떤 request 없이 특정조건에서 자동으로 생성되는 Field 값
+
+    - TypeDefs에 정의한 Type명과 동일하게 resolvers에 함수를 만들면됨
+      받아오는 인자 root는 Computed Field가 실행 전에 실행된 resolver의
+      리턴 값이다.
+      예를 들어 seeProfile을 실행하면 return되는 값은 user 정보이며,
+      Computed Field가 생성되는 시점에 root는 user정보를 가지고 있고
+      해당 user정보로 totalFollowing, totalFollower등의 값을 구할 수 있다.
+
+      Computed Field 실행 순서)
+      GQL 요청 -> seeProfile 실행 -> user 정보 리턴
+      -> root(user) 정보를 인자로 받음 -> Computed Field 값 생성(User Reolver 작동)
+
+# 12/27
+
+1. Hashtag Parasing(#가 붙은 단어만 추출)
+   - 정규식을 이용한다.
+     예시) #[\w]+